Fix android & implement for ios
