+++ android/src/main/java/com/github/rmtmckenzie/qrmobilevision/QrDetector.java
@@ -1,22 +1,17 @@
 package com.github.rmtmckenzie.qrmobilevision;
 
-import android.content.Context;
 import android.util.Log;
-import android.view.Display;
-import android.view.Surface;
-import android.view.WindowManager;
 
 import androidx.annotation.GuardedBy;
 import androidx.annotation.NonNull;
 
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.ml.vision.FirebaseVision;
-import com.google.firebase.ml.vision.barcode.FirebaseVisionBarcode;
-import com.google.firebase.ml.vision.barcode.FirebaseVisionBarcodeDetector;
-import com.google.firebase.ml.vision.barcode.FirebaseVisionBarcodeDetectorOptions;
-import com.google.firebase.ml.vision.common.FirebaseVisionImage;
-import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata;
+import com.google.mlkit.vision.barcode.BarcodeScanner;
+import com.google.mlkit.vision.barcode.BarcodeScannerOptions;
+import com.google.mlkit.vision.barcode.BarcodeScanning;
+import com.google.mlkit.vision.barcode.common.Barcode;
+import com.google.mlkit.vision.common.InputImage;
 
 import java.util.List;
 
@@ -24,13 +19,13 @@ import java.util.List;
  * Allows QrCamera classes to send frames to a Detector
  */
 
-class QrDetector implements OnSuccessListener<List<FirebaseVisionBarcode>>, OnFailureListener {
+class QrDetector implements OnSuccessListener<List<Barcode>>, OnFailureListener {
     private static final String TAG = "cgr.qrmv.QrDetector";
     private final QrReaderCallbacks communicator;
-    private final FirebaseVisionBarcodeDetector detector;
+    private final BarcodeScanner detector;
 
     public interface Frame {
-        FirebaseVisionImage toImage();
+        InputImage toImage();
 
         void close();
     }
@@ -41,9 +36,9 @@ class QrDetector implements OnSuccessListener<List<FirebaseVisionBarcode>>, OnFa
     @GuardedBy("this")
     private Frame processingFrame;
 
-    QrDetector(QrReaderCallbacks communicator, FirebaseVisionBarcodeDetectorOptions options) {
+    QrDetector(QrReaderCallbacks communicator, BarcodeScannerOptions options) {
         this.communicator = communicator;
-        this.detector = FirebaseVision.getInstance().getVisionBarcodeDetector(options);
+        this.detector = BarcodeScanning.getClient(options);
     }
 
     void detect(Frame frame) {
@@ -65,7 +60,7 @@ class QrDetector implements OnSuccessListener<List<FirebaseVisionBarcode>>, OnFa
     }
 
     private void processFrame(Frame frame) {
-        FirebaseVisionImage image;
+        InputImage image;
         try {
             image = frame.toImage();
         } catch (IllegalStateException ex) {
@@ -74,16 +69,17 @@ class QrDetector implements OnSuccessListener<List<FirebaseVisionBarcode>>, OnFa
             return;
         }
 
-        if (image != null)
-        detector.detectInImage(image)
-            .addOnSuccessListener(this)
-            .addOnFailureListener(this);
+        if (image != null) {
+            detector.process(image)
+                .addOnSuccessListener(this)
+                .addOnFailureListener(this);
+        }
         frame.close();
     }
 
     @Override
-    public void onSuccess(List<FirebaseVisionBarcode> firebaseVisionBarcodes) {
-        for (FirebaseVisionBarcode barcode : firebaseVisionBarcodes) {
+    public void onSuccess(List<Barcode> firebaseVisionBarcodes) {
+        for (Barcode barcode : firebaseVisionBarcodes) {
             communicator.qrRead(barcode.getRawValue());
         }
         processLatest();
