+++ lib/src/barcode_formats.dart
@@ -1,0 +1,20 @@
+enum BarcodeFormats {
+  ALL_FORMATS,
+  AZTEC,
+  CODE_128,
+  CODE_39,
+  CODE_93,
+  CODABAR,
+  DATA_MATRIX,
+  EAN_13,
+  EAN_8,
+  ITF,
+  PDF417,
+  QR_CODE,
+  UPC_A,
+  UPC_E,
+}
+
+const defaultBarcodeFormats = const [
+  BarcodeFormats.ALL_FORMATS,
+];
